pneumonia_detector/requirements.txt
pyautogui
pillow
tkcap
pydicom
new_imagepdf
opencv_python
matplotlib
pandas
tensorflow
python-xlib
pytest
